Add contracts command for compiling and deploying smart contracts
Currently this wraps the solidity compiler but in the future it can
support a more pluggble, modular architecture (i.e., eWASM, Viper, or
even non-EVM based smart contracting languages). Also, the current
implementation only supports a single flattened solidity contract (i.e.,
a flattened contract may have any number of libraries or other contracts
defined). The .sol file must also be defined in the source as it is
considered the top-level contract and the target for a deployment to the
provide platform. The architecture of this CLI has been built
thoughtfully enough that supporting standard truffle projects should be
fairly straightforward.